Move synchronisation from observers to cronjobs
The synchronisation was set to run using observers. This causes that
when the synchronisation lacks, the entire flow in Magento is slowed
down.

Another feature that is now possible, is that the actual payment
status can now be send to Smile.io.

Both the customers and the orders are synced to Smile.io when the
new database column "smileio_synchronised_at" column value is
lower then the value of the "updated_at" database column.